Fix line handling, gap position update, and escape sequences
- Corrected type casting in TB_MergeGap() to avoid overflow
- Fixed TB_DeleteCurrentLine() to delete the correct line and maintain line_count
- Prevented TB_LoadFromFile() from deleting the only line in buffer
- Improved File_ReadLine() to strip both '\n' and '\r' endings
- Corrected escape sequence mappings for Shift+Arrow keys (added trailing '~')

+++ src/document/textbuffer.c
@@ -59,7 +59,7 @@ void TB_MergeGap(TextBuffer *tb) {
     UTF8String_Deinit(&a);
     UTF8String_Deinit(&b);
     // set the cursor position to end of the former gap
-    tb->gap.position += tb->gap.text.length - tb->gap.overlap;
+    tb->gap.position += (long)tb->gap.text.length - (long)tb->gap.overlap;
     // reset the gap
     tb->gap.text.length = 0;
     tb->gap.overlap = 0;
